# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AlertStatus } from "./alertStatus";
import type { AlertType } from "./alertType";

export interface Alert {
  id: string;
  /** Alert message */
  text: string;
  /** Human-readable timestamp */
  timestamp: string;
  /** ISO datetime when alert was issued */
  issuedAt: Date;
  status: AlertStatus;
  type: AlertType;
}