# lib/api-zod/src/generated/types/alertsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from "./alert";

export interface AlertsResponse {
  alerts: Alert[];
  total: number;
  activeCount: number;
}