# lib/api-zod/src/generated/types/marketState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketStateBias } from "./marketStateBias";
import type { MarketStateGamma } from "./marketStateGamma";

/**
 * Complete market intelligence state
 */
export interface MarketState {
  /** Directional market bias */
  bias: MarketStateBias;
  /** Net gamma positioning state */
  gamma: MarketStateGamma;
  /** Current key price zone (e.g. "$82K") */
  zone: string;
  /** Active market scenario label */
  scenario: string;
  /** Active setup description (e.g. "RECHAZO → CONTINUACIÓN") */
  setup: string;
  /**
   * Scenario probability percentage
   * @minimum 0
   * @maximum 100
   */
  probability: number;
  /** Directional outlook summary */
  outlook?: string;
  /** Relevant timeframe for the scenario */
  timeframe?: string;
  /** Scenario classification tags */
  tags?: string[];
  /**
   * Bias conviction percentage
   * @minimum 0
   * @maximum 100
   */
  biasStrength?: number;
  /**
   * Gamma level (-100 = max short, +100 = max long)
   * @minimum -100
   * @maximum 100
   */
  gammaLevel?: number;
  /** Net gamma dollar value (e.g. "-$1.2B") */
  netGamma?: string;
  /** Gamma flip price level */
  flipPoint?: string;
  /** Dominant options expiry date */
  dominantExpiry?: string;
  /** ISO timestamp of last update */
  lastUpdate: string;
}