# lib/api-zod/src/generated/types/terminalPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from "./alert";
import type { MarketState } from "./marketState";

/**
 * Full state update from the trading terminal
 */
export interface TerminalPayload {
  marketState: MarketState;
  /** Optional list of new alerts to merge */
  alerts?: Alert[];
}